CRAYSAT-1206: Add guidance for API timeouts for CAPMC during power off
This change adds a note to the "Power Off Compute and IO Cabinets"
section which describes how to resolve issues which may occur when
running the `sat bootsys shutdown --stage cabinet-power` command. In the
event that the command times out, it is now suggested that the command
be run with a higher value for `--api-timeout`.

(cherry picked from commit bb02c93fe9fc8007371e22681fb867858f120952)